Create a new settings panel that contains all app settings
